added mappers folder

+++ DTO/Conditions.cs
@@ -9,7 +9,7 @@ namespace DTO
     public class Conditions
     {
         public Conditions() { }
-        public int comission { get; set; }
+        public double comission { get; set; }
         public double yearlyInterest { get; set; }
         //interes adicional
         public double addYearlyInterest { get; set;}
